update all of the files referencing clipboard copy
