+++ lib/primer/view_components/linters/clipboard_copy_component_migration_counter.rb
@@ -12,9 +12,9 @@ module ERBLint
 
       TAGS = %w[clipboard-copy].freeze
       REQUIRED_ARGUMENTS = [/for|value/, "aria-label"].freeze
-      MESSAGE = "We are migrating clipboard-copy to use [Primer::ClipboardCopy](https://primer.style/view-components/components/clipboardcopy), please try to use that instead of raw HTML."
+      MESSAGE = "We are migrating clipboard-copy to use [Primer::Beta::ClipboardCopy](https://primer.style/view-components/components/clipboardcopy), please try to use that instead of raw HTML."
       ARGUMENT_MAPPER = ArgumentMappers::ClipboardCopy
-      COMPONENT = "Primer::ClipboardCopy"
+      COMPONENT = "Primer::Beta::ClipboardCopy"
     end
   end
 end
