+++ lib/tasks/docs.rake
@@ -53,7 +53,7 @@ namespace :docs do
       Primer::ButtonComponent,
       Primer::Beta::ButtonGroup,
       Primer::Alpha::ButtonMarketing,
-      Primer::ClipboardCopy,
+      Primer::Beta::ClipboardCopy,
       Primer::Beta::CloseButton,
       Primer::Beta::Counter,
       Primer::Beta::Details,
@@ -104,7 +104,7 @@ namespace :docs do
       Primer::Alpha::ImageCrop,
       Primer::Beta::AutoComplete,
       Primer::Alpha::Banner,
-      Primer::ClipboardCopy,
+      Primer::Beta::ClipboardCopy,
       Primer::TabContainerComponent,
       Primer::TimeAgoComponent,
       Primer::Alpha::UnderlinePanels,
