+++ previews/primer/clipboard_copy_preview/element.html.erb
@@ -1,2 +1,2 @@
-<%= render(Primer::ClipboardCopy.new(for: "blob-path", "aria-label": aria_label)) %>
+<%= render(Primer::Beta::ClipboardCopy.new(for: "blob-path", "aria-label": aria_label)) %>
 <div id="blob-path">src/index.js</div>
