+++ static/audited_at.json
@@ -45,6 +45,7 @@
   "Primer::Beta::Breadcrumbs::Item": "",
   "Primer::Beta::Button": "",
   "Primer::Beta::ButtonGroup": "",
+  "Primer::Beta::ClipboardCopy": "",
   "Primer::Beta::CloseButton": "",
   "Primer::Beta::Counter": "",
   "Primer::Beta::Details": "",
