+++ static/constants.json
@@ -558,6 +558,8 @@
   },
   "Primer::Beta::ButtonGroup": {
   },
+  "Primer::Beta::ClipboardCopy": {
+  },
   "Primer::Beta::CloseButton": {
     "DEFAULT_TYPE": "button",
     "TYPE_OPTIONS": [
