+++ static/statuses.json
@@ -45,6 +45,7 @@
   "Primer::Beta::Breadcrumbs::Item": "alpha",
   "Primer::Beta::Button": "beta",
   "Primer::Beta::ButtonGroup": "beta",
+  "Primer::Beta::ClipboardCopy": "beta",
   "Primer::Beta::CloseButton": "beta",
   "Primer::Beta::Counter": "beta",
   "Primer::Beta::Details": "beta",
@@ -62,7 +63,7 @@
   "Primer::Box": "stable",
   "Primer::BoxComponent": "deprecated",
   "Primer::ButtonComponent": "deprecated",
-  "Primer::ClipboardCopy": "beta",
+  "Primer::ClipboardCopy": "deprecated",
   "Primer::ConditionalWrapper": "alpha",
   "Primer::Content": "stable",
   "Primer::Dropdown": "deprecated",
