+++ test/components/beta/clipboard_copy_test.rb
@@ -2,11 +2,11 @@
 
 require "components/test_helper"
 
-class PrimerClipboardCopyTest < Minitest::Test
+class PrimerBetaClipboardCopyTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_renders_simple
-    render_inline Primer::ClipboardCopy.new(value: "my-branch-name", "aria-label": "Copy branch name to clipboard")
+    render_inline Primer::Beta::ClipboardCopy.new(value: "my-branch-name", "aria-label": "Copy branch name to clipboard")
 
     assert_selector("clipboard-copy[data-view-component][value=\"my-branch-name\"]") do
       assert_selector("svg.octicon.octicon-copy")
@@ -16,14 +16,14 @@ class PrimerClipboardCopyTest < Minitest::Test
 
   def test_requires_aria_label_when_empty
     assert_raises(ArgumentError) do
-      render_inline Primer::ClipboardCopy.new(value: "my-branch-name") do
+      render_inline Primer::Beta::ClipboardCopy.new(value: "my-branch-name") do
         nil
       end
     end
   end
 
   def test_renders_with_text_contents
-    render_inline Primer::ClipboardCopy.new(value: "my-branch-name") do
+    render_inline Primer::Beta::ClipboardCopy.new(value: "my-branch-name") do
       "Click to copy!"
     end
 
@@ -33,7 +33,7 @@ class PrimerClipboardCopyTest < Minitest::Test
   end
 
   def test_renders_with_for
-    render_inline Primer::ClipboardCopy.new(for: "element-id", "aria-label": "Copy branch name to clipboard")
+    render_inline Primer::Beta::ClipboardCopy.new(for: "element-id", "aria-label": "Copy branch name to clipboard")
 
     assert_selector("clipboard-copy[data-view-component][for=\"element-id\"]") do
       assert_selector("svg.octicon.octicon-copy")
